requirements.txt
pdfplumber
pytesseract
Pillow
flask-cors
